transform coordinator tests into sensor tests
