+++ tests/components/libre_hardware_monitor/test_sensor.py
@@ -2,17 +2,27 @@
 
 from dataclasses import replace
 from datetime import timedelta
+import logging
 from types import MappingProxyType
-from unittest.mock import AsyncMock
+from unittest.mock import AsyncMock, patch
 
 from freezegun.api import FrozenDateTimeFactory
 from librehardwaremonitor_api import LibreHardwareMonitorConnectionError
+from librehardwaremonitor_api.model import (
+    DeviceId,
+    DeviceName,
+    LibreHardwareMonitorData,
+)
+import pytest
 from syrupy.assertion import SnapshotAssertion
 
-from homeassistant.components.libre_hardware_monitor.const import DEFAULT_SCAN_INTERVAL
+from homeassistant.components.libre_hardware_monitor.const import (
+    DEFAULT_SCAN_INTERVAL,
+    DOMAIN,
+)
 from homeassistant.const import STATE_UNAVAILABLE, STATE_UNKNOWN
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers import entity_registry as er
+from homeassistant.helpers import device_registry as dr, entity_registry as er
 
 from . import init_integration
 
@@ -134,3 +144,62 @@ async def test_sensor_state_is_unknown_when_no_sensor_data_is_provided(
 
     assert state
     assert state.state == STATE_UNKNOWN
+
+
+async def test_orphaned_devices_are_removed(
+    hass: HomeAssistant,
+    mock_lhm_client: AsyncMock,
+    mock_config_entry: MockConfigEntry,
+    freezer: FrozenDateTimeFactory,
+) -> None:
+    """Test that devices in HA that do not receive updates are removed."""
+    await init_integration(hass, mock_config_entry)
+
+    mock_lhm_client.get_data.return_value = LibreHardwareMonitorData(
+        main_device_ids_and_names=MappingProxyType(
+            {
+                DeviceId("amdcpu-0"): DeviceName("AMD Ryzen 7 7800X3D"),
+                DeviceId("gpu-nvidia-0"): DeviceName("NVIDIA GeForce RTX 4080 SUPER"),
+            }
+        ),
+        sensor_data=mock_lhm_client.get_data.return_value.sensor_data,
+    )
+
+    device_registry = dr.async_get(hass)
+    orphaned_device = device_registry.async_get_or_create(
+        config_entry_id=mock_config_entry.entry_id,
+        identifiers={(DOMAIN, "lpc-nct6687d-0")},
+    )
+
+    with patch.object(
+        device_registry,
+        "async_remove_device",
+        wraps=device_registry.async_update_device,
+    ) as mock_remove:
+        freezer.tick(timedelta(DEFAULT_SCAN_INTERVAL))
+        async_fire_time_changed(hass)
+        await hass.async_block_till_done()
+
+        mock_remove.assert_called_once_with(orphaned_device.id)
+
+
+async def test_integration_does_not_log_new_devices_on_first_refresh(
+    hass: HomeAssistant,
+    mock_lhm_client: AsyncMock,
+    mock_config_entry: MockConfigEntry,
+    caplog: pytest.LogCaptureFixture,
+) -> None:
+    """Test that initial data update does not cause warning about new devices."""
+    mock_lhm_client.get_data.return_value = LibreHardwareMonitorData(
+        main_device_ids_and_names=MappingProxyType(
+            {
+                **mock_lhm_client.get_data.return_value.main_device_ids_and_names,
+                DeviceId("generic-memory"): DeviceName("Generic Memory"),
+            }
+        ),
+        sensor_data=mock_lhm_client.get_data.return_value.sensor_data,
+    )
+
+    with caplog.at_level(logging.WARNING):
+        await init_integration(hass, mock_config_entry)
+        assert len(caplog.records) == 0
